Fix build break on HP-UX and other platforms
(lwreg: r37613)

+++ lwreg/shell/cmdparse.c
@@ -378,7 +378,7 @@ RegShellHexStringExtentLen(
 
     for (dwIndx=0; pszHexString[dwIndx]; dwIndx++)
     {
-        if (!isxdigit(pszHexString[dwIndx]))
+        if (!isxdigit((int) pszHexString[dwIndx]))
         {
             break;
         }
